replace storage/cassandra
