+++ build.gradle
@@ -25,7 +25,6 @@ configurations {
 }
 
 dependencies {
-    compile group: 'com.datastax.cassandra', name: 'cassandra-driver-core', version: '3.6.0'
     compile group: 'com.datastax.oss', name: 'java-driver-core', version: '4.3.0'
     compile group: 'com.datastax.oss', name: 'java-driver-query-builder', version: '4.3.0'
     compile group: 'com.google.guava', name: 'guava', version: '24.1-jre'
