+++ src/main/java/com/scalar/database/storage/cassandra/ConditionSetter.java
@@ -1,12 +1,12 @@
 package com.scalar.database.storage.cassandra;
 
-import static com.datastax.driver.core.querybuilder.QueryBuilder.*;
+import static com.datastax.oss.driver.api.querybuilder.QueryBuilder.bindMarker;
 
-import com.datastax.driver.core.querybuilder.BuiltStatement;
-import com.datastax.driver.core.querybuilder.Clause;
-import com.datastax.driver.core.querybuilder.Delete;
-import com.datastax.driver.core.querybuilder.Insert;
-import com.datastax.driver.core.querybuilder.Update;
+import com.datastax.oss.driver.api.querybuilder.BuildableQuery;
+import com.datastax.oss.driver.api.querybuilder.condition.Condition;
+import com.datastax.oss.driver.api.querybuilder.delete.Delete;
+import com.datastax.oss.driver.api.querybuilder.insert.Insert;
+import com.datastax.oss.driver.api.querybuilder.update.Update;
 import com.scalar.database.api.ConditionalExpression;
 import com.scalar.database.api.DeleteIf;
 import com.scalar.database.api.DeleteIfExists;
@@ -14,6 +14,7 @@ import com.scalar.database.api.MutationConditionVisitor;
 import com.scalar.database.api.PutIf;
 import com.scalar.database.api.PutIfExists;
 import com.scalar.database.api.PutIfNotExists;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.IntStream;
 import javax.annotation.concurrent.NotThreadSafe;
@@ -21,19 +22,23 @@ import javax.annotation.concurrent.NotThreadSafe;
 /**
  * A visitor class to configure condition-specific statement
  *
- * @author Hiroyuki Yamada
+ * @author Hiroyuki Yamada, Yuji Ito
  */
 @NotThreadSafe
 public class ConditionSetter implements MutationConditionVisitor {
-  private final BuiltStatement statement;
+  private BuildableQuery query;
 
   /**
-   * Constructs {@code ConditionSetter} with the specified {@code BuiltStatement}
+   * Constructs {@code ConditionSetter} with the specified {@code BuildableQuery}
    *
-   * @param statement {@code BuiltStatement} to set conditions
+   * @param query {@code BuildableQuery} to set conditions
    */
-  public ConditionSetter(BuiltStatement statement) {
-    this.statement = statement;
+  public ConditionSetter(BuildableQuery query) {
+    this.query = query;
+  }
+
+  public BuildableQuery getQuery() {
+    return query;
   }
 
   /**
@@ -43,15 +48,15 @@ public class ConditionSetter implements MutationConditionVisitor {
    */
   @Override
   public void visit(PutIf condition) {
-    Update.Where update = (Update.Where) statement;
-
     List<ConditionalExpression> expressions = condition.getExpressions();
-    Update.Conditions cond = update.onlyIf(createClauseWith(expressions.get(0)));
-    IntStream.range(1, expressions.size())
+    List<Condition> conditions = new ArrayList<>();
+    IntStream.range(0, expressions.size())
         .forEach(
             i -> {
-              cond.and(createClauseWith(expressions.get(i)));
+              conditions.add(createConditionWith(expressions.get(i)));
             });
+
+    query = ((Update) query).if_(conditions);
   }
 
   /**
@@ -61,8 +66,7 @@ public class ConditionSetter implements MutationConditionVisitor {
    */
   @Override
   public void visit(PutIfExists condition) {
-    Update.Where update = (Update.Where) statement;
-    update.ifExists();
+    query = ((Update) query).ifExists();
   }
 
   /**
@@ -72,8 +76,7 @@ public class ConditionSetter implements MutationConditionVisitor {
    */
   @Override
   public void visit(PutIfNotExists condition) {
-    Insert insert = (Insert) statement;
-    insert.ifNotExists();
+    query = ((Insert) query).ifNotExists();
   }
 
   /**
@@ -83,15 +86,15 @@ public class ConditionSetter implements MutationConditionVisitor {
    */
   @Override
   public void visit(DeleteIf condition) {
-    Delete.Where delete = (Delete.Where) statement;
-
     List<ConditionalExpression> expressions = condition.getExpressions();
-    Delete.Conditions cond = delete.onlyIf(createClauseWith(expressions.get(0)));
-    IntStream.range(1, expressions.size())
+    List<Condition> conditions = new ArrayList<>();
+    IntStream.range(0, expressions.size())
         .forEach(
             i -> {
-              cond.and(createClauseWith(expressions.get(i)));
+              conditions.add(createConditionWith(expressions.get(i)));
             });
+
+    query = ((Delete) query).if_(conditions);
   }
 
   /**
@@ -101,24 +104,23 @@ public class ConditionSetter implements MutationConditionVisitor {
    */
   @Override
   public void visit(DeleteIfExists condition) {
-    Delete.Where delete = (Delete.Where) statement;
-    delete.ifExists();
+    query = ((Delete) query).ifExists();
   }
 
-  private Clause createClauseWith(ConditionalExpression e) {
+  private Condition createConditionWith(ConditionalExpression e) {
     switch (e.getOperator()) {
       case EQ:
-        return eq(e.getName(), bindMarker());
+        return Condition.column(e.getName()).isEqualTo(bindMarker());
       case NE:
-        return ne(e.getName(), bindMarker());
+        return Condition.column(e.getName()).isNotEqualTo(bindMarker());
       case GT:
-        return gt(e.getName(), bindMarker());
+        return Condition.column(e.getName()).isGreaterThan(bindMarker());
       case GTE:
-        return gte(e.getName(), bindMarker());
+        return Condition.column(e.getName()).isGreaterThanOrEqualTo(bindMarker());
       case LT:
-        return lt(e.getName(), bindMarker());
+        return Condition.column(e.getName()).isLessThan(bindMarker());
       case LTE:
-        return lte(e.getName(), bindMarker());
+        return Condition.column(e.getName()).isLessThanOrEqualTo(bindMarker());
       default:
         // never comes here because ConditionalExpression accepts only above operators
         throw new IllegalArgumentException(e.getOperator() + " is not supported");
