+++ src/main/java/com/scalar/database/storage/cassandra/DeleteStatementHandler.java
@@ -1,54 +1,59 @@
 package com.scalar.database.storage.cassandra;
 
-import static com.datastax.driver.core.querybuilder.QueryBuilder.bindMarker;
+import static com.datastax.oss.driver.api.querybuilder.QueryBuilder.bindMarker;
+import static com.datastax.oss.driver.api.querybuilder.QueryBuilder.deleteFrom;
 
-import com.datastax.driver.core.BoundStatement;
-import com.datastax.driver.core.PreparedStatement;
-import com.datastax.driver.core.ResultSet;
-import com.datastax.driver.core.Session;
-import com.datastax.driver.core.querybuilder.QueryBuilder;
-import com.scalar.database.api.Delete;
+import com.datastax.oss.driver.api.core.CqlSession;
+import com.datastax.oss.driver.api.core.cql.BoundStatement;
+import com.datastax.oss.driver.api.core.cql.BoundStatementBuilder;
+import com.datastax.oss.driver.api.core.cql.PreparedStatement;
+import com.datastax.oss.driver.api.core.cql.ResultSet;
+import com.datastax.oss.driver.api.querybuilder.delete.Delete;
+import com.datastax.oss.driver.api.querybuilder.delete.DeleteSelection;
+import com.datastax.oss.driver.api.querybuilder.relation.Relation;
 import com.scalar.database.api.Operation;
+import java.util.ArrayList;
+import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.concurrent.ThreadSafe;
 
 /**
  * A handler class for delete statements
  *
- * @author Hiroyuki Yamada
+ * @author Hiroyuki Yamada, Yuji Ito
  */
 @ThreadSafe
 public class DeleteStatementHandler extends MutateStatementHandler {
 
   /**
-   * Constructs a {@code DeleteStatementHandler} with the specified {@code Session}
+   * Constructs a {@code DeleteStatementHandler} with the specified {@code CqlSession}
    *
    * @param session session to be used with this statement
    */
-  public DeleteStatementHandler(Session session) {
+  public DeleteStatementHandler(CqlSession session) {
     super(session);
   }
 
   @Override
   @Nonnull
   protected PreparedStatement prepare(Operation operation) {
-    checkArgument(operation, Delete.class);
+    checkArgument(operation, com.scalar.database.api.Delete.class);
 
-    Delete del = (Delete) operation;
-    com.datastax.driver.core.querybuilder.Delete delete = prepare(del);
+    com.scalar.database.api.Delete del = (com.scalar.database.api.Delete) operation;
+    Delete delete = prepare(del);
 
-    return prepare(delete.getQueryString());
+    return prepare(delete.asCql());
   }
 
   @Override
   @Nonnull
-  protected BoundStatement bind(PreparedStatement prepared, Operation operation) {
-    checkArgument(operation, Delete.class);
+  protected BoundStatementBuilder bind(PreparedStatement prepared, Operation operation) {
+    checkArgument(operation, com.scalar.database.api.Delete.class);
 
-    BoundStatement bound = prepared.bind();
-    bound = bind(bound, (Delete) operation);
+    BoundStatementBuilder builder = prepared.boundStatementBuilder();
+    bind(builder, (com.scalar.database.api.Delete) operation);
 
-    return bound;
+    return builder;
   }
 
   @Override
@@ -57,32 +62,28 @@ public class DeleteStatementHandler extends MutateStatementHandler {
     return session.execute(bound);
   }
 
-  private com.datastax.driver.core.querybuilder.Delete prepare(Delete del) {
-    com.datastax.driver.core.querybuilder.Delete delete =
-        QueryBuilder.delete().from(del.forNamespace().get(), del.forTable().get());
-    com.datastax.driver.core.querybuilder.Delete.Where where = delete.where();
+  private Delete prepare(com.scalar.database.api.Delete del) {
+    DeleteSelection delete = deleteFrom(del.forNamespace().get(), del.forTable().get());
 
-    del.getPartitionKey().forEach(v -> where.and(QueryBuilder.eq(v.getName(), bindMarker())));
+    List<Relation> relations = new ArrayList<Relation>();
+    del.getPartitionKey()
+        .forEach(v -> relations.add(Relation.column(v.getName()).isEqualTo(bindMarker())));
     del.getClusteringKey()
         .ifPresent(
             k -> {
-              k.forEach(v -> where.and(QueryBuilder.eq(v.getName(), bindMarker())));
+              k.forEach(v -> relations.add(Relation.column(v.getName()).isEqualTo(bindMarker())));
             });
 
-    setCondition(where, del);
-
-    return delete;
+    return (Delete) setCondition(delete.where(relations), del);
   }
 
-  private BoundStatement bind(BoundStatement bound, Delete del) {
-    ValueBinder binder = new ValueBinder(bound);
+  private void bind(BoundStatementBuilder builder, com.scalar.database.api.Delete del) {
+    ValueBinder binder = new ValueBinder(builder);
 
     // bind in the prepared order
     del.getPartitionKey().forEach(v -> v.accept(binder));
     del.getClusteringKey().ifPresent(k -> k.forEach(v -> v.accept(binder)));
 
     bindCondition(binder, del);
-
-    return bound;
   }
 }
