+++ src/main/java/com/scalar/database/storage/cassandra/ScannerIterator.java
@@ -1,7 +1,7 @@
 package com.scalar.database.storage.cassandra;
 
-import com.datastax.driver.core.ResultSet;
-import com.datastax.driver.core.Row;
+import com.datastax.oss.driver.api.core.cql.ResultSet;
+import com.datastax.oss.driver.api.core.cql.Row;
 import com.scalar.database.api.Result;
 import java.util.Iterator;
 import javax.annotation.Nullable;
