+++ src/main/java/com/scalar/database/storage/cassandra/StatementHandler.java
@@ -2,11 +2,12 @@ package com.scalar.database.storage.cassandra;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.datastax.driver.core.BoundStatement;
-import com.datastax.driver.core.ConsistencyLevel;
-import com.datastax.driver.core.PreparedStatement;
-import com.datastax.driver.core.ResultSet;
-import com.datastax.driver.core.Session;
+import com.datastax.oss.driver.api.core.ConsistencyLevel;
+import com.datastax.oss.driver.api.core.cql.BoundStatement;
+import com.datastax.oss.driver.api.core.cql.BoundStatementBuilder;
+import com.datastax.oss.driver.api.core.cql.PreparedStatement;
+import com.datastax.oss.driver.api.core.cql.ResultSet;
+import com.datastax.oss.driver.api.core.CqlSession;
 import com.google.common.base.Joiner;
 import com.scalar.database.api.Consistency;
 import com.scalar.database.api.Operation;
@@ -21,18 +22,15 @@ import org.slf4j.LoggerFactory;
 @ThreadSafe
 public abstract class StatementHandler {
   private static final Logger LOGGER = LoggerFactory.getLogger(StatementHandler.class);
-  protected final Session session;
-  protected StatementCache cache;
+  protected final CqlSession session;
 
   /**
-   * Constructs a {@code StatementHandler} with the specified {@link Session} and a new {@link
-   * StatementCache}
+   * Constructs a {@code StatementHandler} with the specified {@link CqlSession}
    *
-   * @param session {@code Session}
+   * @param session {@code CqlSession}
    */
-  protected StatementHandler(Session session) {
+  protected StatementHandler(CqlSession session) {
     this.session = checkNotNull(session);
-    cache = new StatementCache();
   }
 
   /**
@@ -65,9 +63,9 @@ public abstract class StatementHandler {
   @Nonnull
   protected ResultSet handleInternal(Operation operation) {
     PreparedStatement prepared = prepare(operation);
-    BoundStatement bound = bind(prepared, operation);
-    setConsistency(bound, operation);
-    return execute(bound, operation);
+    BoundStatementBuilder builder = bind(prepared, operation);
+    setConsistency(builder, operation);
+    return execute(builder.build(), operation);
   }
 
   /**
@@ -79,35 +77,28 @@ public abstract class StatementHandler {
   @Nonnull
   protected PreparedStatement prepare(String queryString) {
     LOGGER.debug("query to prepare : [" + queryString + "].");
-    PreparedStatement prepared = cache.get(queryString);
-    if (prepared == null) {
-      prepared = session.prepare(queryString);
-      cache.put(queryString, prepared);
-    } else {
-      LOGGER.debug("there was a hit in the statement cache for [" + queryString + "].");
-    }
-    return prepared;
+    return session.prepare(queryString);
   }
 
   /**
-   * Sets the consistency level for the specified {@link BoundStatement} and {@link Operation}
+   * Sets the consistency level for the specified {@link BoundStatementBuilder} and {@link Operation}
    *
-   * @param bound a {@code BoundStatement}
+   * @param builder a {@code BoundStatementBuilder}
    * @param operation an {@code Operation}
    */
-  protected void setConsistency(BoundStatement bound, Operation operation) {
-    bound.setConsistencyLevel(StatementHandler.convert(operation, operation.getConsistency()));
+  protected void setConsistency(BoundStatementBuilder builder, Operation operation) {
+    builder.setConsistencyLevel(StatementHandler.convert(operation, operation.getConsistency()));
     // set preferable consistency for the operation
-    overwriteConsistency(bound, operation);
+    overwriteConsistency(builder, operation);
   }
 
   protected abstract PreparedStatement prepare(Operation operation);
 
-  protected abstract BoundStatement bind(PreparedStatement prepared, Operation operation);
+  protected abstract BoundStatementBuilder bind(PreparedStatement prepared, Operation operation);
 
   protected abstract ResultSet execute(BoundStatement bound, Operation operation);
 
-  protected abstract void overwriteConsistency(BoundStatement bound, Operation operation);
+  protected abstract void overwriteConsistency(BoundStatementBuilder builder, Operation operation);
 
   /**
    * Returns a {@link ConsistencyLevel} based on the specified {@link Operation} and {@link
