+++ src/main/java/com/scalar/database/storage/cassandra/ValueBinder.java
@@ -2,7 +2,7 @@ package com.scalar.database.storage.cassandra;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.datastax.driver.core.BoundStatement;
+import com.datastax.oss.driver.api.core.cql.BoundStatementBuilder;
 import com.scalar.database.io.BigIntValue;
 import com.scalar.database.io.BlobValue;
 import com.scalar.database.io.BooleanValue;
@@ -17,95 +17,95 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * A visitor class to bind {@code Value}s to a {@link BoundStatement}
+ * A visitor class to bind {@code Value}s to a {@link BoundStatementBuilder}
  *
- * @author Hiroyuki Yamada
+ * @author Hiroyuki Yamada, Yuji Ito
  */
 @NotThreadSafe
 public final class ValueBinder implements ValueVisitor {
   private static final Logger LOGGER = LoggerFactory.getLogger(ValueBinder.class);
-  private final BoundStatement bound;
+  private final BoundStatementBuilder builder;
   private int i;
 
   /**
-   * Constructs {@code ValueBinder} with the specified {@code BoundStatement}
+   * Constructs {@code ValueBinder} with the specified {@code BoundStatementBuilder}
    *
-   * @param bound a {@code BoundStatement} to be bound
+   * @param builder a {@code BoundStatementBuilder} to be bound
    */
-  public ValueBinder(BoundStatement bound) {
-    this.bound = checkNotNull(bound);
+  public ValueBinder(BoundStatementBuilder builder) {
+    this.builder = checkNotNull(builder);
     i = 0;
   }
 
   /**
-   * Sets the specified {@code BooleanValue} to the bound statement
+   * Sets the specified {@code BooleanValue} to the bound statement builder
    *
    * @param value a {@code BooleanValue} to be set
    */
   @Override
   public void visit(BooleanValue value) {
     LOGGER.debug(value.get() + " is bound to " + i);
-    bound.setBool(i++, value.get());
+    builder.setBoolean(i++, value.get());
   }
 
   /**
-   * Sets the specified {@code IntValue} to the bound statement
+   * Sets the specified {@code IntValue} to the bound statement builder
    *
    * @param value a {@code IntValue} to be set
    */
   @Override
   public void visit(IntValue value) {
     LOGGER.debug(value.get() + " is bound to " + i);
-    bound.setInt(i++, value.get());
+    builder.setInt(i++, value.get());
   }
 
   /**
-   * Sets the specified {@code BigIntValue} to the bound statement
+   * Sets the specified {@code BigIntValue} to the bound statement builder
    *
    * @param value a {@code BigIntValue} to be set
    */
   @Override
   public void visit(BigIntValue value) {
     LOGGER.debug(value.get() + " is bound to " + i);
-    bound.setLong(i++, value.get());
+    builder.setLong(i++, value.get());
   }
 
   /**
-   * Sets the specified {@code FloatValue} to the bound statement
+   * Sets the specified {@code FloatValue} to the bound statement builder
    *
    * @param value a {@code FloatValue} to be set
    */
   @Override
   public void visit(FloatValue value) {
     LOGGER.debug(value.get() + " is bound to " + i);
-    bound.setFloat(i++, value.get());
+    builder.setFloat(i++, value.get());
   }
 
   /**
-   * Sets the specified {@code DoubleValue} to the bound statement
+   * Sets the specified {@code DoubleValue} to the bound statement builder
    *
    * @param value a {@code DoubleValue} to be set
    */
   @Override
   public void visit(DoubleValue value) {
     LOGGER.debug(value.get() + " is bound to " + i);
-    bound.setDouble(i++, value.get());
+    builder.setDouble(i++, value.get());
   }
 
   /**
-   * Sets the specified {@code TextValue} to the bound statement
+   * Sets the specified {@code TextValue} to the bound statement builder
    *
    * @param value a {@code TextValue} to be set
    */
   @Override
   public void visit(TextValue value) {
     LOGGER.debug(value.getString() + " is bound to " + i);
-    value.getString().ifPresent(s -> bound.setString(i, s));
+    value.getString().ifPresent(s -> builder.setString(i, s));
     i++;
   }
 
   /**
-   * Sets the specified {@code BlobValue} to the bound statement
+   * Sets the specified {@code BlobValue} to the bound statement builder
    *
    * @param value a {@code BlobValue} to be set
    */
@@ -116,7 +116,7 @@ public final class ValueBinder implements ValueVisitor {
         .get()
         .ifPresent(
             b -> {
-              bound.setBytes(i, (ByteBuffer) ByteBuffer.allocate(b.length).put(b).flip());
+              builder.setByteBuffer(i, (ByteBuffer) ByteBuffer.allocate(b.length).put(b).flip());
             });
     i++;
   }
