--- src/main/java/com/scalar/database/storage/cassandra4driver/ScannerImpl.java
@@ -1,52 +1,0 @@
-package com.scalar.database.storage.cassandra4driver;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-import com.datastax.oss.driver.api.core.cql.ResultSet;
-import com.datastax.oss.driver.api.core.cql.Row;
-import com.scalar.database.api.Result;
-import com.scalar.database.api.Scanner;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Optional;
-import javax.annotation.Nonnull;
-import javax.annotation.concurrent.NotThreadSafe;
-
-@NotThreadSafe
-public final class ScannerImpl implements Scanner {
-  private final ResultSet resultSet;
-  private final TableMetadata metadata;
-
-  public ScannerImpl(ResultSet resultSet, TableMetadata metadata) {
-    this.resultSet = checkNotNull(resultSet);
-    this.metadata = checkNotNull(metadata);
-  }
-
-  @Override
-  @Nonnull
-  public Optional<Result> one() {
-    Row row = resultSet.one();
-    if (row == null) {
-      return Optional.empty();
-    }
-    return Optional.of(new ResultImpl(row, metadata));
-  }
-
-  @Override
-  @Nonnull
-  public List<Result> all() {
-    List<Result> results = new ArrayList<>();
-    resultSet.forEach(r -> results.add(new ResultImpl(r, metadata)));
-    return results;
-  }
-
-  @Override
-  public Iterator<Result> iterator() {
-    return new ScannerIterator(resultSet, metadata);
-  }
-
-  @Override
-  public void close() {
-  }
-}
