+++ src/test/java/com/scalar/database/storage/cassandra/ClusterManagerTest.java
@@ -7,19 +7,20 @@ import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.KeyspaceMetadata;
-import com.datastax.driver.core.Metadata;
-import com.datastax.driver.core.Session;
-import com.datastax.driver.core.TableMetadata;
-import com.datastax.driver.core.exceptions.DriverException;
-import com.datastax.driver.core.exceptions.NoHostAvailableException;
+import com.datastax.oss.driver.api.core.CqlSession;
+import com.datastax.oss.driver.api.core.CqlSessionBuilder;
+import com.datastax.oss.driver.api.core.DriverException;
+import com.datastax.oss.driver.api.core.metadata.Metadata;
+import com.datastax.oss.driver.api.core.metadata.schema.KeyspaceMetadata;
+import com.datastax.oss.driver.api.core.metadata.schema.TableMetadata;
 import com.scalar.database.config.DatabaseConfig;
 import com.scalar.database.exception.storage.ConnectionException;
 import com.scalar.database.exception.storage.StorageRuntimeException;
+import java.util.Optional;
 import java.util.Properties;
 import org.junit.Before;
 import org.junit.Test;
@@ -31,12 +32,13 @@ import org.mockito.MockitoAnnotations;
 public class ClusterManagerTest {
   private static final String VALID_ADDRESS = "127.0.0.1";
   private static final String INVALID_ADDRESS = "1024.0.0.1";
+  private static final String INVALID_PORT = Integer.toString(Integer.MAX_VALUE);
   private static final String ANY_USERNAME = "username";
   private static final String ANY_PASSWORD = "password";
   private static final String ANY_KEYSPACE_NAME = "keyspace";
   private static final String ANY_TABLE_NAME = "table";
-  @Mock private Cluster cluster;
-  @Mock private Session session;
+  @Mock private CqlSession session;
+  @Mock private CqlSessionBuilder builder;
   private ClusterManager manager;
   private Properties props;
   private DatabaseConfig config;
@@ -52,8 +54,8 @@ public class ClusterManagerTest {
     config = new DatabaseConfig(props);
 
     manager = Mockito.spy(new ClusterManager(config));
-    doReturn(cluster).when(manager).getCluster(config);
-    when(cluster.connect()).thenReturn(session);
+    doReturn(builder).when(manager).getBuilder();
+    when(builder.build()).thenReturn(session);
   }
 
   @Test
@@ -66,8 +68,8 @@ public class ClusterManagerTest {
         .doesNotThrowAnyException();
 
     // Assert
-    verify(manager).build();
-    verify(cluster).connect();
+    verify(manager).getBuilder();
+    verify(builder).build();
   }
 
   @Test
@@ -82,18 +84,36 @@ public class ClusterManagerTest {
             () -> {
               manager.getSession();
             })
+        .isInstanceOf(ConnectionException.class);
+
+    // Assert
+    verify(builder, never()).build();
+  }
+
+  @Test
+  public void getSession_ConfigWithInvalidPortGiven_ShouldThrowExceptionProperly() {
+    // Arrange
+    props.setProperty(DatabaseConfig.CONTACT_PORT, INVALID_PORT);
+    config = new DatabaseConfig(props);
+    manager = Mockito.spy(new ClusterManager(config));
+
+    // Act Assert
+    assertThatThrownBy(
+            () -> {
+              manager.getSession();
+            })
         .isInstanceOf(ConnectionException.class)
         .hasCauseExactlyInstanceOf(IllegalArgumentException.class);
 
     // Assert
-    verify(cluster, never()).connect();
+    verify(builder, never()).build();
   }
 
   @Test
   public void getSession_CalledOnceAndDriverExceptionThrown_ShouldThrowConnectionException() {
     // Arrange
     DriverException toThrow = mock(DriverException.class);
-    when(cluster.connect()).thenThrow(toThrow);
+    when(builder.build()).thenThrow(toThrow);
 
     // Act Assert
     assertThatThrownBy(
@@ -104,14 +124,15 @@ public class ClusterManagerTest {
         .hasCause(toThrow);
 
     // Assert
-    verify(manager).build();
-    verify(cluster).connect();
+    verify(manager).getBuilder();
+    verify(builder).build();
   }
 
   @Test
-  public void getSession_CalledAfterBuilt_ShouldNotBuild() {
+  public void getSession_CalledAfterClosed_ShouldBuildAgain() {
     // Arrange
-    manager.build();
+    manager.getSession();
+    manager.close();
 
     // Act Assert
     assertThatCode(
@@ -121,8 +142,8 @@ public class ClusterManagerTest {
         .doesNotThrowAnyException();
 
     // Assert
-    verify(manager).build();
-    verify(cluster).connect();
+    verify(manager, times(2)).getBuilder();
+    verify(builder, times(2)).build();
   }
 
   @Test
@@ -138,8 +159,8 @@ public class ClusterManagerTest {
         .doesNotThrowAnyException();
 
     // Assert
-    verify(manager).build();
-    verify(cluster).connect();
+    verify(manager, times(1)).getBuilder();
+    verify(builder, times(1)).build();
   }
 
   @Test
@@ -159,9 +180,9 @@ public class ClusterManagerTest {
     Metadata metadata = mock(Metadata.class);
     KeyspaceMetadata keyspaceMetadata = mock(KeyspaceMetadata.class);
     TableMetadata tableMetadata = mock(TableMetadata.class);
-    when(cluster.getMetadata()).thenReturn(metadata);
-    when(metadata.getKeyspace(anyString())).thenReturn(keyspaceMetadata);
-    when(keyspaceMetadata.getTable(anyString())).thenReturn(tableMetadata);
+    when(session.getMetadata()).thenReturn(metadata);
+    when(metadata.getKeyspace(anyString())).thenReturn(Optional.of(keyspaceMetadata));
+    when(keyspaceMetadata.getTable(anyString())).thenReturn(Optional.of(tableMetadata));
 
     // Act
     TableMetadata actual = manager.getMetadata(ANY_KEYSPACE_NAME, ANY_TABLE_NAME);
@@ -171,28 +192,12 @@ public class ClusterManagerTest {
   }
 
   @Test
-  public void getMetadata_NoHostAvailable_ShouldThrowConnectionException() {
-    // Arrange
-    manager.getSession();
-    NoHostAvailableException toThrow = mock(NoHostAvailableException.class);
-    when(cluster.getMetadata()).thenThrow(toThrow);
-
-    // Act Assert
-    assertThatThrownBy(
-            () -> {
-              manager.getMetadata(ANY_KEYSPACE_NAME, ANY_TABLE_NAME);
-            })
-        .isInstanceOf(ConnectionException.class)
-        .hasCause(toThrow);
-  }
-
-  @Test
   public void getMetadata_KeyspaceNotExists_ShouldThrowStorageRuntimeException() {
     // Arrange
     manager.getSession();
     Metadata metadata = mock(Metadata.class);
-    when(cluster.getMetadata()).thenReturn(metadata);
-    when(metadata.getKeyspace(anyString())).thenReturn(null);
+    when(session.getMetadata()).thenReturn(metadata);
+    when(metadata.getKeyspace(anyString())).thenReturn(Optional.empty());
 
     // Act Assert
     assertThatThrownBy(
@@ -208,9 +213,9 @@ public class ClusterManagerTest {
     manager.getSession();
     Metadata metadata = mock(Metadata.class);
     KeyspaceMetadata keyspaceMetadata = mock(KeyspaceMetadata.class);
-    when(cluster.getMetadata()).thenReturn(metadata);
-    when(metadata.getKeyspace(anyString())).thenReturn(keyspaceMetadata);
-    when(keyspaceMetadata.getTable(anyString())).thenReturn(null);
+    when(session.getMetadata()).thenReturn(metadata);
+    when(metadata.getKeyspace(anyString())).thenReturn(Optional.of(keyspaceMetadata));
+    when(keyspaceMetadata.getTable(anyString())).thenReturn(Optional.empty());
 
     // Act
     assertThatThrownBy(
