+++ src/test/java/com/scalar/database/storage/cassandra/UpdateStatementHandlerTest.java
@@ -8,10 +8,10 @@ import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.datastax.driver.core.BoundStatement;
-import com.datastax.driver.core.ConsistencyLevel;
-import com.datastax.driver.core.PreparedStatement;
-import com.datastax.driver.core.Session;
+import com.datastax.oss.driver.api.core.ConsistencyLevel;
+import com.datastax.oss.driver.api.core.CqlSession;
+import com.datastax.oss.driver.api.core.cql.BoundStatementBuilder;
+import com.datastax.oss.driver.api.core.cql.PreparedStatement;
 import com.google.common.base.Joiner;
 import com.scalar.database.api.ConditionalExpression;
 import com.scalar.database.api.Consistency;
@@ -29,8 +29,8 @@ import org.mockito.MockitoAnnotations;
 
 /** */
 public class UpdateStatementHandlerTest {
-  private static final String ANY_KEYSPACE_NAME = "keyspace";
-  private static final String ANY_TABLE_NAME = "table";
+  private static final String ANY_KEYSPACE_NAME = "ks";
+  private static final String ANY_TABLE_NAME = "tbl";
   private static final String ANY_NAME_1 = "name1";
   private static final String ANY_NAME_2 = "name2";
   private static final String ANY_NAME_3 = "name3";
@@ -43,9 +43,9 @@ public class UpdateStatementHandlerTest {
   private static final int ANY_INT_2 = 2;
   private UpdateStatementHandler handler;
   private Put put;
-  @Mock private Session session;
+  @Mock private CqlSession session;
   @Mock private PreparedStatement prepared;
-  @Mock private BoundStatement bound;
+  @Mock private BoundStatementBuilder builder;
 
   @Before
   public void setUp() throws Exception {
@@ -79,10 +79,10 @@ public class UpdateStatementHandlerTest {
   private void configureBehavior(String expected) {
     when(session.prepare(expected == null ? anyString() : expected)).thenReturn(prepared);
 
-    when(prepared.bind()).thenReturn(bound);
-    when(bound.setString(anyInt(), anyString())).thenReturn(bound);
-    when(bound.setInt(anyInt(), anyInt())).thenReturn(bound);
-    when(bound.setConsistencyLevel(any(ConsistencyLevel.class))).thenReturn(bound);
+    when(prepared.boundStatementBuilder()).thenReturn(builder);
+    when(builder.setString(anyInt(), anyString())).thenReturn(builder);
+    when(builder.setInt(anyInt(), anyInt())).thenReturn(builder);
+    when(builder.setConsistencyLevel(any(ConsistencyLevel.class))).thenReturn(builder);
   }
 
   @Test
@@ -96,40 +96,14 @@ public class UpdateStatementHandlerTest {
                   "UPDATE",
                   ANY_KEYSPACE_NAME + "." + ANY_TABLE_NAME,
                   "SET",
-                  ANY_NAME_2 + "=?," + ANY_NAME_3 + "=?",
+                  ANY_NAME_2 + "=?, " + ANY_NAME_3 + "=?",
                   "WHERE",
-                  ANY_NAME_1 + "=?;"
+                  ANY_NAME_1 + "=?"
                 });
     configureBehavior(expected);
     put = preparePut();
 
     // Act
-    handler.prepare(put);
-
-    // Assert
-    verify(session).prepare(expected);
-  }
-
-  @Test
-  public void prepare_SameQueryGivenTwice_SecondTimeShouldUseStatementCache() {
-    // Arrange
-    String expected =
-        Joiner.on(" ")
-            .skipNulls()
-            .join(
-                new String[] {
-                  "UPDATE",
-                  ANY_KEYSPACE_NAME + "." + ANY_TABLE_NAME,
-                  "SET",
-                  ANY_NAME_2 + "=?," + ANY_NAME_3 + "=?",
-                  "WHERE",
-                  ANY_NAME_1 + "=?;"
-                });
-    configureBehavior(expected);
-    put = preparePut();
-
-    // Act
-    handler.prepare(put);
     handler.prepare(put);
 
     // Assert
@@ -151,7 +125,7 @@ public class UpdateStatementHandlerTest {
                   "WHERE",
                   ANY_NAME_1 + "=?",
                   "AND",
-                  ANY_NAME_2 + "=?;"
+                  ANY_NAME_2 + "=?"
                 });
     configureBehavior(expected);
     put = preparePutWithClusteringKey();
@@ -179,7 +153,7 @@ public class UpdateStatementHandlerTest {
                   ANY_NAME_1 + "=?",
                   "AND",
                   ANY_NAME_2 + "=?",
-                  "IF EXISTS;"
+                  "IF EXISTS"
                 });
     configureBehavior(expected);
     put = preparePutWithClusteringKey();
@@ -219,7 +193,7 @@ public class UpdateStatementHandlerTest {
                   "AND",
                   ANY_NAME_4 + "<?",
                   "AND",
-                  ANY_NAME_4 + "<=?;"
+                  ANY_NAME_4 + "<=?"
                 });
     configureBehavior(expected);
     put = preparePutWithClusteringKey();
@@ -281,9 +255,9 @@ public class UpdateStatementHandlerTest {
     handler.bind(prepared, put);
 
     // Assert
-    verify(bound).setInt(0, ANY_INT_1);
-    verify(bound).setString(1, ANY_TEXT_1);
-    verify(bound).setString(2, ANY_TEXT_2);
+    verify(builder).setInt(0, ANY_INT_1);
+    verify(builder).setString(1, ANY_TEXT_1);
+    verify(builder).setString(2, ANY_TEXT_2);
   }
 
   @Test
@@ -300,11 +274,11 @@ public class UpdateStatementHandlerTest {
     handler.bind(prepared, put);
 
     // Assert
-    verify(bound).setInt(0, ANY_INT_1);
-    verify(bound).setString(1, ANY_TEXT_1);
-    verify(bound).setString(2, ANY_TEXT_2);
-    verify(bound).setInt(3, ANY_INT_2);
-    verify(bound).setString(4, ANY_TEXT_3);
+    verify(builder).setInt(0, ANY_INT_1);
+    verify(builder).setString(1, ANY_TEXT_1);
+    verify(builder).setString(2, ANY_TEXT_2);
+    verify(builder).setInt(3, ANY_INT_2);
+    verify(builder).setString(4, ANY_TEXT_3);
   }
 
   @Test
@@ -315,10 +289,10 @@ public class UpdateStatementHandlerTest {
     put.withConsistency(Consistency.SEQUENTIAL);
 
     // Act
-    handler.setConsistency(bound, put);
+    handler.setConsistency(builder, put);
 
     // Assert
-    verify(bound).setConsistencyLevel(ConsistencyLevel.QUORUM);
+    verify(builder).setConsistencyLevel(ConsistencyLevel.QUORUM);
   }
 
   @Test
@@ -329,10 +303,10 @@ public class UpdateStatementHandlerTest {
     put.withConsistency(Consistency.EVENTUAL);
 
     // Act
-    handler.setConsistency(bound, put);
+    handler.setConsistency(builder, put);
 
     // Assert
-    verify(bound).setConsistencyLevel(ConsistencyLevel.ONE);
+    verify(builder).setConsistencyLevel(ConsistencyLevel.ONE);
   }
 
   @Test
@@ -344,10 +318,10 @@ public class UpdateStatementHandlerTest {
     put.withConsistency(Consistency.LINEARIZABLE);
 
     // Act
-    handler.setConsistency(bound, put);
+    handler.setConsistency(builder, put);
 
     // Assert
-    verify(bound).setConsistencyLevel(ConsistencyLevel.QUORUM);
+    verify(builder).setConsistencyLevel(ConsistencyLevel.QUORUM);
   }
 
   @Test
@@ -358,11 +332,11 @@ public class UpdateStatementHandlerTest {
     put.withCondition(new PutIfExists()).withConsistency(Consistency.EVENTUAL);
 
     // Act
-    handler.setConsistency(bound, put);
+    handler.setConsistency(builder, put);
 
     // Assert
-    verify(bound).setConsistencyLevel(ConsistencyLevel.QUORUM);
-    verify(bound).setSerialConsistencyLevel(ConsistencyLevel.SERIAL);
+    verify(builder).setConsistencyLevel(ConsistencyLevel.QUORUM);
+    verify(builder).setSerialConsistencyLevel(ConsistencyLevel.SERIAL);
   }
 
   @Test
@@ -375,11 +349,11 @@ public class UpdateStatementHandlerTest {
         .withConsistency(Consistency.EVENTUAL);
 
     // Act
-    handler.setConsistency(bound, put);
+    handler.setConsistency(builder, put);
 
     // Assert
-    verify(bound).setConsistencyLevel(ConsistencyLevel.QUORUM);
-    verify(bound).setSerialConsistencyLevel(ConsistencyLevel.SERIAL);
+    verify(builder).setConsistencyLevel(ConsistencyLevel.QUORUM);
+    verify(builder).setSerialConsistencyLevel(ConsistencyLevel.SERIAL);
   }
 
   /** Unit testing for handle() method is covered in InsertStatementHandlerTest */
